refactor: HowToPlayModalをコンポーネントとして分離、src/app/page.tsxの単一責任化
